Feat: Add UI for categories page 1. Create subcategory tab for category page 2. Update dummy json data 3. Funding slide bug fix
