+++ src/pages/Category.tsx
@@ -1,5 +1,5 @@
 import Header from "../components/header/Header";
-import SubItemCategory from "../components/main/SubItemCategory";
+import SubItemCategory from "../components/category/SubItemCategory";
 import Footer from "../components/footer/Footer";
 
 export default function Category() {
